redux-persist added

+++ frontend/src/components/Product.tsx
@@ -3,11 +3,9 @@ import { Product } from "../types/Product";
 import { deleteProduct } from "../store/product/productActions";
 import Card from "react-bootstrap/Card";
 import Button from "react-bootstrap/Button";
-import ButtonGroup from "react-bootstrap/ButtonGroup";
 import { useNavigate } from "react-router-dom";
 import { useAppDispatch } from "../hooks/redux-hooks";
 
-//{ product, showProductDetail, showEffectDetail }
 type Props = {
   product: Product;
 };
@@ -15,7 +13,7 @@ type Props = {
 const SingleProduct = ({ product }: Props) => {
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
-  
+
   const handleOnClick = () => {
     //showProductDetail(product);
     console.log("product");
@@ -43,15 +41,10 @@ const SingleProduct = ({ product }: Props) => {
   // };
 
   return (
-    <Card
-      className={styles.info}
-      border="success"
-      onClick={() => console.log("product")}
-    >
+    <Card className={styles.info} border="success">
       <Card.Img variant="top">
         {/* <img src={`/files/${product.image}`} alt={product.name} /> */}
       </Card.Img>
-      {/* {showEffectDetail && <h2 className={styles.shadow}>DETAIL</h2>} */}
       <Card.Body className={styles.data}>
         <Card.Title>{product.name}</Card.Title>
         <Card.Title>Price: {product.price} €</Card.Title>
@@ -61,14 +54,22 @@ const SingleProduct = ({ product }: Props) => {
         <Card.Text>{product.description}</Card.Text>
       </Card.Body>
       <Card.Footer className={styles.buttonGroup}>
-        <Button variant="success" onClick={() => navigate(`/edit-product/${product.id}`)}>
+        <Button
+          variant="success"
+          onClick={() => navigate(`/edit-product/${product.id}`)}
+        >
           <i className="fas fa-pencil-alt"></i>
         </Button>
-        <Button variant="danger" onClick={() => {
-          if(product.id) {
-            dispatch(deleteProduct(product.id))
-          }
-        }}>
+        <Button
+          variant="danger"
+          onClick={() => {
+            if (window.confirm("Are you sure to delete?")) {
+              if (product.id) {
+                dispatch(deleteProduct(product.id));
+              }
+            }
+          }}
+        >
           <i className="fas fa-trash-alt"></i>
         </Button>
       </Card.Footer>
